Adds marker to the map without layer
We found out a way to add a layer with a pointer and it worked fine in
e.g. http://geojson.io but for apparently for some styling reasons the
markes itself did not appear on the map.
This solution bypasses layers and current styles and just (forcefully)
adds a marker element to the coordinates.