Makes init-poc-graph-with-holodata add the needed holochain entries
Does not implement hc-entity or similar because the return value is needed twice,
first when adding the node (gowood asset) and second when adding the composed-of edge.
The solution here is to first create the entries and then use address in both cases.
There are other ways to deal with this but this once is sufficient for now.
Also it would be trivial to add the trees to holochain as well.