Dockerizes frontend and adds it to the docker-compose services
There were problems running npm install in the container if there was a
volume mount for the /frontend directory.
Therefore the contents are now copied to the container in the build
phase.
This works but is not perfect in a sense that if one wants to update the
frontend, the frontend image needs to removed manually before running
docker-compose again. On the other hand, in development mode it's
probably best to run the frontend locally anyway.